+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The following is a curated list of changes in the Enact project, newest changes 
 	functionality, only markup.
 - Spotlight containers to spot the last focused element by default.
 - `ui/Group` prop `select` to `childSelect` and added prop `select` to support selection types
+- `ui/ViewManager` added `preventUpdate` prop.
 
 ### Removed
 
@@ -50,6 +51,7 @@ The following is a curated list of changes in the Enact project, newest changes 
 
 - Spotlight stops at container boundaries when 5-way key held down
 - Several issues related to spotting controls in edge cases
+- Joined picker now has correct animation onMouseWheel. 
 
 
 ## [1.0.0-alpha.2] - 2016-10-21